src: Store token in Header object to avoid additional lookups

+++ src/nghttp.h
@@ -122,11 +122,11 @@ struct Request {
 
   bool is_ipv6_literal_addr() const;
 
-  bool response_pseudo_header_allowed(int token) const;
-  bool push_request_pseudo_header_allowed(int token) const;
+  bool response_pseudo_header_allowed(int16_t token) const;
+  bool push_request_pseudo_header_allowed(int16_t token) const;
 
-  Headers::value_type *get_res_header(int token);
-  Headers::value_type *get_req_header(int token);
+  Headers::value_type *get_res_header(int16_t token);
+  Headers::value_type *get_req_header(int16_t token);
 
   void record_request_time();
   void record_response_time();
